Move sync phase logic from SyncAdapter into per-phase controllers.
SyncAdapter and DbSyncHelper are two huge classes with lots of static methods that can be better
isolated from each other. There also isn't always a clear delineation of responsibility between
the two classes - sometimes code in DbSyncHelper makes network requests and sometimes it doesn't,
for example.

In the interest of making this code more maintainable, this commit moves the various `updateFoo`
methods into separate `SyncPhaseController` classes. Each sync phase now has its own
`SyncPhaseController` that's responsible for executing that phase, including the network fetch,
any processing, and the database update.
